# crates/server/src/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: DumbMonit HTTP API
  version: "0.1"
  summary: Monitoring for homelabs and small teams — the API the web UI uses, open to scripts and assistants.
  description: |
    Everything the DumbMonit web interface does goes through this API. It is an
    **alpha**: routes and fields may still change between releases. The
    `X-DumbMonit-Api-Version` header, present on every `/api` response, says
    which revision of this document the server implements.

    ## Authentication

    Two ways in, accepted on every protected route:

    - **API token** — `Authorization: Bearer dmt_…`, created in Settings →
      *API & assistants*. Meant for scripts, dashboards and assistants. No
      cookie is involved, so no anti-CSRF header is needed.
    - **Session cookie** — `dumbmonit_session`, what the web UI uses after
      `POST /api/auth/login`. Every state-changing request (`POST`, `PUT`,
      `DELETE`) sent with the cookie must also carry
      `X-Requested-With: DumbMonit`, or it is refused with `403`.

    A token has a scope. **`read`** grants what a *viewer* account sees: every
    `GET`. **`write`** grants what an *administrator* does: `POST`, `PUT` and
    `DELETE` as well. Whatever its scope, a token can never manage accounts,
    sessions, two-factor authentication, SSO, API or agent tokens, nor export
    or restore a backup: those operations are marked `x-token-scope: none` and
...
    `read` one.

...
    `read`, `write` or `none` (session only).

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/auth/setup:
    post:
      operationId: setupFirstAdmin
      tags: [Accounts]
      summary: Create the first administrator
      description: Only on a fresh instance (`409` once an account exists), with the setup code printed in the server logs (`401` if wrong, `429` after too many attempts). Does not open a session.
      security: []
      x-token-scope: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "204": { description: Account created. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/auth/login:
    post:
      operationId: login
      tags: [Accounts]
      summary: Sign in with a password
      description: >-
        `204` with a `Set-Cookie` header carrying the session. When the account
        has two-factor authentication, `200` with a pending ticket to finish
        with `POST /api/auth/login/totp`. Rate limited after five failures
        (`429`, `Retry-After` doubling from 30 s to 5 min).
      security: []
      x-token-scope: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "204":
          description: Signed in; the session cookie is set.
          headers:
            Set-Cookie:
              description: "`dumbmonit_session`, HttpOnly, SameSite=Lax."
              schema: { type: string }
        "200":
          description: Two-factor authentication required.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TotpChallenge" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/auth/login/totp:
    post:
      operationId: loginTotp
      tags: [Accounts]
      summary: Finish a sign-in with a two-factor code
      security: []
      x-token-scope: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TotpLoginRequest" }
      responses:
        "204":
          description: Signed in; the session cookie is set.
          headers:
            Set-Cookie:
              description: "`dumbmonit_session`, HttpOnly, SameSite=Lax."
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Accounts]
      summary: End the current session
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      responses:
        "204": { description: Signed out; the cookie is cleared. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Accounts]
      summary: The signed-in account
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/auth/password:
    post:
      operationId: changePassword
      tags: [Accounts]
      summary: Change the password of the signed-in account
      description: Signs out every other session of the account.
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChange" }
      responses:
        "204": { description: Password changed. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: dmt_…
      description: >-
        API token created in Settings → API & assistants. `read` scope: every
        `GET`; `write` scope: also `POST`, `PUT`, `DELETE`. Never valid on
        operations marked `x-token-scope: none`. 120 calls per minute per token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: dumbmonit_session
      description: >-
        Session opened by `POST /api/auth/login`. State-changing requests must
        also carry `X-Requested-With: DumbMonit`.
...
  headers:
    RetryAfter:
      description: Seconds to wait before trying again.
      schema: { type: integer }
    RateLimitLimit:
      description: Calls allowed per minute for this token.
      schema: { type: integer }
    RateLimitRemaining:
      description: Calls left in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until the window resets.
      schema: { type: integer }
    WwwAuthenticate:
      description: "`Bearer` (with `error=\"invalid_token\"` for an expired token)."
      schema: { type: string }
  responses:
    BadRequest:
      description: The request is malformed or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No valid session or API token (unknown, revoked or expired token).
      headers:
        WWW-Authenticate: { $ref: "#/components/headers/WwwAuthenticate" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >-
        Not allowed: viewer or `read` token on a write operation, token on a
        session-only operation or from a network it is not allowed from,
        missing `X-Requested-With: DumbMonit` on a cookie write.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: Not possible in the current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limited.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable message. }
...
    Role:
      type: string
      enum: [admin, viewer]
    User:
      type: object
      required: [id, username, display_name, role, auth, disabled, totp_enabled, created_at]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        display_name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        auth: { type: string, enum: [password, oidc] }
        disabled: { type: boolean }
        totp_enabled: { type: boolean }
        created_at: { type: string }
        last_login_at: { type: [string, "null"] }
...
    SetupRequest:
      type: object
      required: [setup_code, password]
      properties:
        setup_code:
          type: string
          writeOnly: true
          description: One-time code printed in the server logs at startup (or `DUMBMONIT_SETUP_CODE`). Case, dashes and spaces are ignored.
        username: { type: string, default: admin }
        password: { type: string, minLength: 12, writeOnly: true }
    LoginRequest:
      type: object
      required: [password]
      properties:
        username: { type: string }
        password: { type: string, writeOnly: true }
    TotpChallenge:
      type: object
      required: [totp_required, pending]
      properties:
        totp_required: { type: boolean, const: true }
        pending: { type: string, description: 'Ticket valid five minutes, dead after five wrong codes.' }
    TotpLoginRequest:
      type: object
      required: [pending, code]
      properties:
        pending: { type: string }
        code: { type: string, description: A six-digit code or a recovery code. }
    PasswordChange:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, writeOnly: true }
        new_password: { type: string, writeOnly: true }
...